implemented Exception handling

+++ task-manager-backend/src/main/java/com/swathika/taskmanager/repository/TaskRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.swathika.taskmanager.entities.Task;
 
 
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
+
+	
 
 }
